+++ pshell/procs.py
@@ -7,7 +7,7 @@ import psutil
 from .env import resolve_env
 
 
-__all__ = ('find_procs_by_cmdline', 'kill_procs')
+__all__ = ('find_procs_by_cmdline', 'kill')
 
 
 def find_procs_by_cmdline(*cmdlines):
@@ -54,8 +54,9 @@ def find_procs_by_cmdline(*cmdlines):
 
     procs = []
     for proc in psutil.process_iter():
+        print("pid=", proc.pid)
         try:
-            # Process.username() *always* fails on Windows with AccessDenied
+            # On Windows, proc.username() ALWAYS fails
             if os.name != 'nt' and proc.username() != getpass.getuser():
                 continue
             cmdline = " ".join(proc.cmdline())
@@ -76,8 +77,8 @@ def find_procs_by_cmdline(*cmdlines):
     return procs
 
 
-def kill_procs(procs, *, term_timeout=10):
-    """Send SIGTERM to a list of processes. After ``term_timeout`` seconds,
+def kill(*procs, term_timeout=10):
+    """Send SIGTERM to one or more processes. After ``term_timeout`` seconds,
     send SIGKILL to the surviving processes.
 
     This function will return before ``term_timeout`` if all processes close
@@ -88,9 +89,8 @@ def kill_procs(procs, *, term_timeout=10):
     current process and its parents.
 
     :param procs:
-        sequence of :class:`psutil.Process` objects, e.g. as returned by
-        :func:`find_procs_by_cmdline`.
-
+        one or more PIDs (int) or :class:`psutil.Process` objects, e.g. as
+        returned by :func:`find_procs_by_cmdline`.
     :param float term_timeout:
         seconds to wait between SIGTERM and SIGKILL.
         If ``term_timeout==0``, skip SIGTERM and immediately send SIGKILL.
@@ -99,6 +99,21 @@ def kill_procs(procs, *, term_timeout=10):
     new_procs = []
     my_pid = os.getpid()
     for proc in procs:
+        # Convert any int PIDs to psutil.Process
+        if isinstance(proc, int):
+            try:
+                proc = psutil.Process(proc)
+            except psutil.NoSuchProcess:
+                logging.debug("PID %d does not exist", proc)
+                continue
+        elif proc is None:
+            # Silently skip - useful as e.g. psutil.Process.parent() can
+            # return None
+            continue
+        elif not isinstance(proc, psutil.Process):
+            raise TypeError("Expected int or psutil.Process; got %s" %
+                            type(proc))
+
         try:
             if proc.pid == my_pid:
                 logging.debug("Not terminating PID %d as it is the current "
@@ -114,7 +129,8 @@ def kill_procs(procs, *, term_timeout=10):
             continue
 
         new_procs.append(proc)
-    procs = new_procs
+
+    procs, new_procs = new_procs, []
 
     if term_timeout != 0 and procs:
         logging.info("Sending SIGTERM to PIDs %s",
@@ -122,12 +138,16 @@ def kill_procs(procs, *, term_timeout=10):
         for proc in procs:
             try:
                 proc.terminate()
+                new_procs.append(proc)
             except psutil.NoSuchProcess:
                 # Process already died
                 pass
+            except psutil.AccessDenied:
+                logging.info("Failed to send SIGTERM to PID %d: access denied",
+                             proc.pid)
 
         # Wait up to <term_timeout> seconds for SIGTERM to be received
-        _, procs = psutil.wait_procs(procs, term_timeout)
+        _, procs = psutil.wait_procs(new_procs, term_timeout)
 
     if procs:
         logging.info("Sending SIGKILL to PIDs %s",
@@ -138,5 +158,17 @@ def kill_procs(procs, *, term_timeout=10):
             except psutil.NoSuchProcess:
                 # Process already died
                 pass
+            except psutil.AccessDenied:
+                logging.info("Failed to send SIGKILL to PID %d: access denied",
+                             proc.pid)
 
     logging.info("All processes terminated")
+
+
+def killall(*cmdlines, term_timeout=10):
+    """Find all processes with the target command line(s), send SIGTERM, and
+    then send SIGKILL to the survivors.
+
+    See :func:`find_procs_by_cmdline` and :func:`kill`.
+    """
+    kill(*find_procs_by_cmdline(*cmdlines), term_timeout=term_timeout)